pyproject.toml
[project]
name = "nonebot-plugin-blocker"
version = "0.3.2"
description = "Message Blocker"
authors = [{ name = "MerCuJerry", email = "mercujerry@gmail.com" }]
keywords = ["nonebot", "nonebot2", "qqbot", "bot"]
dependencies = [
    "nonebot2[fastapi] >= 2.0.0",
    "nonebot-adapter-onebot>=2.1.0",
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = {file = "LICENSE"}

[project.urls]
homepage = "https://github.com/MerCuJerry/nonebot-plugin-blocker"

[tool.pdm.build]
includes = ["nonebot_plugin_blocker"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"